refactor(browser): extract container RPC manager
